requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
pytesseract==0.3.13
Pillow==11.0.0